+++ cynic-proc-macros/src/lib.rs
@@ -4,7 +4,7 @@ use proc_macro::TokenStream;
 
 use cynic_codegen::{
     enum_derive, fragment_arguments_derive, fragment_derive, inline_fragments_derive, query_dsl,
-    query_module, scalars_as_strings,
+    query_module, scalar_derive,
 };
 
 #[proc_macro]
@@ -12,19 +12,6 @@ pub fn query_dsl(input: TokenStream) -> TokenStream {
     let input = syn::parse_macro_input!(input as query_dsl::QueryDslParams);
 
     let rv = query_dsl::query_dsl_from_schema(input).unwrap().into();
-
-    //eprintln!("{}", rv);
-
-    rv
-}
-
-#[proc_macro]
-pub fn scalars_as_strings(input: TokenStream) -> TokenStream {
-    let input = syn::parse_macro_input!(input as scalars_as_strings::Params);
-
-    let rv = scalars_as_strings::scalars_as_strings(input)
-        .unwrap()
-        .into();
 
     //eprintln!("{}", rv);
 
@@ -76,6 +63,20 @@ pub fn enum_derive(input: TokenStream) -> TokenStream {
     let ast = syn::parse_macro_input!(input as syn::DeriveInput);
 
     let rv = match enum_derive::enum_derive(&ast) {
+        Ok(tokens) => tokens.into(),
+        Err(e) => e.to_compile_error().into(),
+    };
+
+    //eprintln!("{}", rv);
+
+    rv
+}
+
+#[proc_macro_derive(Scalar)]
+pub fn scalar_derive(input: TokenStream) -> TokenStream {
+    let ast = syn::parse_macro_input!(input as syn::DeriveInput);
+
+    let rv = match scalar_derive::scalar_derive(&ast) {
         Ok(tokens) => tokens.into(),
         Err(e) => e.to_compile_error().into(),
     };